+++ xsimlab/dot.py
@@ -9,14 +9,11 @@ Part of the code below is copied and modified from:
   http://dask.pydata.org
 
 """
-from __future__ import absolute_import, division, print_function
-
 import os
 from functools import partial
 
-from .utils import import_required
-from .variable.base import (AbstractVariable, ForeignVariable,
-                            DiagnosticVariable, VariableGroup)
+from .utils import attr_fields_dict, import_required, maybe_to_list
+from .variable import VarIntent, VarType
 
 
 graphviz = import_required("graphviz", "Drawing dask graphs requires the "
@@ -36,91 +33,91 @@ VAR_EDGE_ATTRS = {'arrowhead': 'none', 'color': '#555555'}
 
 
 def hash_variable(var):
-    return str(hash(var))
+    # issue with variables with the same name declared in different processes
+    # return str(hash(var))
+    return str(id(var))
 
 
 def _add_processes(g, model):
     seen = set()
 
-    for proc_name, proc in model._processes.items():
-        label = proc_name
-        if proc_name not in seen:
-            seen.add(proc_name)
-            g.node(proc_name, label=label, **PROC_NODE_ATTRS)
+    for p_name, p_obj in model._processes.items():
+        if p_name not in seen:
+            seen.add(p_name)
+            g.node(p_name, label=p_name, **PROC_NODE_ATTRS)
 
-        for dep_proc_name in model._dep_processes[proc_name]:
-            # check and add node shouldn't be needed here, but not sure
-            #if dep_proc_name not in seen:
-            #    seen.add(dep_proc_name)
-            #    dep_label = dep_proc_name
-            #    g.node(dep_proc_name, label=dep_label, **PROC_NODE_ATTRS)
-            g.edge(dep_proc_name, proc_name, **PROC_EDGE_ATTRS)
+        for dep_p_name in model.dependent_processes[p_name]:
+            g.edge(dep_p_name, p_name, **PROC_EDGE_ATTRS)
 
 
-def _add_var(g, var, label, link_2_node, is_input=False):
+def _add_var(g, model, var, p_name):
     node_attrs = VAR_NODE_ATTRS.copy()
     edge_attrs = VAR_EDGE_ATTRS.copy()
-    var_key = hash_variable(var)
 
-    if is_input:
+    var_key = hash_variable(var)
+    var_intent = var.metadata['intent']
+    var_type = var.metadata['var_type']
+
+    if (p_name, var.name) in model._input_vars:
         node_attrs = INPUT_NODE_ATTRS.copy()
         edge_attrs = INPUT_EDGE_ATTRS.copy()
-    elif isinstance(var, DiagnosticVariable):
+    elif var_type == VarType.ON_DEMAND:
         node_attrs['style'] = 'diagonals'
-    elif isinstance(var, ForeignVariable):
+    elif var_type == VarType.FOREIGN:
         node_attrs['style'] = 'dashed'
         edge_attrs['style'] = 'dashed'
-    elif isinstance(var, (tuple, VariableGroup)):
+    elif var_type == VarType.GROUP:
         node_attrs['shape'] = 'box3d'
 
-    if not isinstance(var, tuple) and var.provided:
+    if var_intent == VarIntent.OUT:
         edge_attrs.update({'arrowhead': 'empty'})
-        edge_ends = link_2_node, var_key
+        edge_ends = p_name, var_key
     else:
-        edge_ends = var_key, link_2_node
+        edge_ends = var_key, p_name
 
-    g.node(var_key, label=label, **node_attrs)
+    g.node(var_key, label=var.name, **node_attrs)
     g.edge(*edge_ends, weight='200', **edge_attrs)
 
 
 def _add_inputs(g, model):
-    for proc_name, variables in model._input_vars.items():
-        for var_name, var in variables.items():
-            _add_var(g, var, var_name, proc_name, is_input=True)
+    for p_name, var_name in model._input_vars:
+        p_cls = type(model[p_name])
+        var = attr_fields_dict(p_cls)[var_name]
+
+        _add_var(g, model, var, p_name)
 
 
 def _add_variables(g, model):
-    for proc_name, variables in model._processes.items():
-        for var_name, var in variables.items():
-            if model.is_input(var):
-                continue
-            _add_var(g, var, var_name, proc_name)
+    for p_name, p_obj in model._processes.items():
+        p_cls = type(p_obj)
 
-            if isinstance(var, (tuple, VariableGroup)):
-                for v in var:
-                    _add_var(g, v, '\<no_name\>', hash_variable(var))
+        for var_name, var in attr_fields_dict(p_cls).items():
+            _add_var(g, model, var, p_name)
 
 
-def _add_var_and_foreign_vars(g, model, proc_name, var_name):
-    variable = model[proc_name][var_name]
+def _get_target_keys(p_obj, var_name):
+    return (
+        maybe_to_list(p_obj.__xsimlab_store_keys__.get(var_name, [])) +
+        maybe_to_list(p_obj.__xsimlab_od_keys__.get(var_name, []))
+    )
 
-    if isinstance(variable, ForeignVariable):
-        variable = variable.ref_var
 
-    for p_name, variables in model._processes.items():
-        for v_name, var in variables.items():
-            if model.is_input(var):
-                is_input = True
-            else:
-                is_input = False
+def _add_var_and_targets(g, model, p_name, var_name):
+    this_p_name = p_name
+    this_var_name = var_name
 
-            if var is variable or getattr(var, 'ref_var', None) is variable:
-                _add_var(g, var, v_name, p_name, is_input=is_input)
-            elif isinstance(var, (tuple, VariableGroup)):
-                for v in var:
-                    if v is variable or getattr(v, 'ref_var', None) is variable:
-                        _add_var(g, var, v_name, p_name, is_input=is_input)
-                        _add_var(g, v, '\<no_name\>', hash_variable(var))
+    this_p_obj = model._processes[this_p_name]
+    this_target_keys = _get_target_keys(this_p_obj, this_var_name)
+
+    for p_name, p_obj in model._processes.items():
+        p_cls = type(p_obj)
+
+        for var_name, var in attr_fields_dict(p_cls).items():
+            target_keys = _get_target_keys(p_obj, var_name)
+
+            if ((p_name, var_name) == (this_p_name, this_var_name) or
+                    len(set(target_keys) & set(this_target_keys))):
+                _add_var(g, model, var, p_name)
 
 
 def to_graphviz(model, rankdir='LR', show_only_variable=None,
@@ -134,18 +131,14 @@ def to_graphviz(model, rankdir='LR', show_only_variable=None,
     _add_processes(g, model)
 
     if show_only_variable is not None:
-        if isinstance(show_only_variable, AbstractVariable):
-            proc_name, var_name = model._get_proc_var_name(show_only_variable)
-        else:
-            proc_name, var_name = show_only_variable
-        _add_var_and_foreign_vars(g, model, proc_name, var_name)
+        p_name, var_name = show_only_variable
+        _add_var_and_targets(g, model, p_name, var_name)
 
     else:
-        if show_inputs:
-            _add_inputs(g, model)
-
         if show_variables:
             _add_variables(g, model)
+        elif show_inputs:
+            _add_inputs(g, model)
 
     return g
 
@@ -188,27 +181,25 @@ def dot_graph(model, filename=None, format=None, show_only_variable=None,
               show_inputs=False, show_variables=False, **kwargs):
     """
     Render a model as a graph using dot.
-    If `filename` is not None, write a file to disk with that name in the
-    format specified by `format`.  `filename` should not include an extension.
 
     Parameters
     ----------
     model : object
         The Model instance to display.
     filename : str or None, optional
-        The name (without an extension) of the file to write to disk.  If
+        The name (without an extension) of the file to write to disk. If
         `filename` is None (default), no file will be written, and we
         communicate with dot using only pipes.
     format : {'png', 'pdf', 'dot', 'svg', 'jpeg', 'jpg'}, optional
         Format in which to write output file.  Default is 'png'.
-    show_only_variable : object or tuple, optional
-        Show only a variable (and all other linked variables) given either
-        as a Variable object or a tuple corresponding to process name and
-        variable name. Deactivated by default.
+    show_only_variable : tuple, optional
+        Show only a variable (and all other variables sharing the
+        same value) given as a tuple ``(process_name, variable_name)``.
+        Deactivated by default.
     show_inputs : bool, optional
         If True, show all input variables in the graph (default: False).
         Ignored if `show_only_variable` is not None.
-    show_variabless : bool, optional
+    show_variables : bool, optional
         If True, show also the other variables (default: False).
         Ignored if `show_only_variable` is not None.
     **kwargs
@@ -216,7 +207,8 @@ def dot_graph(model, filename=None, format=None, show_only_variable=None,
 
     Returns
     -------
-    result : None or IPython.display.Image or IPython.display.SVG  (See below.)
+    result : None or IPython.display.Image or IPython.display.SVG
+        (See below.)
 
     Notes
     -----